+++ lib/commands/partner/premium.dart
@@ -131,4 +131,5 @@ final Group premiumCommand = Group(
       },
     ),
   ],
+  aliases: ["vip"],
 );