+++ lib/fishstick_dart.dart
@@ -1,5 +1,7 @@
 import "client/client.dart";
+
 export "commands/info/ping.dart";
+export "commands/info/invite.dart";
 export "commands/partner/premium.dart";
 export "commands/admin/partner.dart";